Seperate Auth and Registration traits.
Simplifies opting out of Registration.
