+++ src/Illuminate/Foundation/Auth/Redirectable.php
@@ -1,0 +1,20 @@
+<?php namespace Illuminate\Foundation\Auth;
+
+trait Redirectable {
+
+	/**
+	 * Get the post register / login redirect path.
+	 *
+	 * @return string
+	 */
+	public function redirectPath()
+	{
+		if (property_exists($this, 'redirectPath'))
+		{
+			return $this->redirectPath;
+		}
+
+		return property_exists($this, 'redirectTo') ? $this->redirectTo : '/home';
+	}
+
+}
