+++ src/Illuminate/Foundation/Auth/RegistersUsers.php
@@ -1,0 +1,42 @@
+<?php namespace Illuminate\Foundation\Auth;
+
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
+trait RegistersUsers {
+
+	use Redirectable;
+
+	/**
+	 * Show the application registration form.
+	 *
+	 * @return \Illuminate\Http\Response
+	 */
+	public function getRegister()
+	{
+		return view('auth.register');
+	}
+
+	/**
+	 * Handle a registration request for the application.
+	 *
+	 * @param  \Illuminate\Http\Request  $request
+	 * @return \Illuminate\Http\Response
+	 */
+	public function postRegister(Request $request)
+	{
+		$validator = $this->validator($request->all());
+
+		if ($validator->fails())
+		{
+			$this->throwValidationException(
+				$request, $validator
+			);
+		}
+
+		Auth::login($this->create($request->all()));
+
+		return redirect($this->redirectPath());
+	}
+
+}
